Deduplicate advantage payload builder into common; mirror snapshot-workflow fixes to slime
_advantage_samples_payload moved to common/reporting.py so miles no longer
imports it from the slime package. The slime snapshot workflow gets the same
fixes the miles one already had: trigger on patcher/recipe changes and stage
untracked snapshot files before the drift diff.
