+++ .github/workflows/validate-slime-patch-snapshots.yml
@@ -5,6 +5,8 @@ on:
     paths:
       - ".github/workflows/validate-slime-patch-snapshots.yml"
       - "modal_training_gym/frameworks/slime/launcher.py"
+      - "modal_training_gym/frameworks/slime/modal_helpers/patches/*.py"
+      - "modal_training_gym/train_recipes/slime_recipe/recipe.py"
       - "scripts/fetch_slime_patch_snapshots.py"
       - "tests/testdata/*.input"
 
@@ -34,4 +36,6 @@ jobs:
         run: uv run modal run scripts/fetch_slime_patch_snapshots.py
 
       - name: Verify committed snapshot is current
-        run: git diff --exit-code -- tests/testdata/*.input
+        run: |
+          git add -A -- tests/testdata/*.input
+          git diff --cached --exit-code -- tests/testdata/*.input
